--- a/pmedian.jl
+++ b/pmedian.jl
@@ -213,6 +213,9 @@ function solve_moi(data::PMedianData, optimizer; params)
         MOI.Utilities.UniversalFallback(MOI.Utilities.Model{Float64}()),
         optimizer())
     model = MOI.Bridges.full_bridge_optimizer(cached, Float64)
+    # resetting optimizer is necessary to use copy_to to transfer all the data
+    # from caching optimizer to GLPK in a single batch. If that is not called
+    # constraints are passed one by one during model generation phase.
     MOI.Utilities.reset_optimizer(cached)
     for (param, value) in params
         MOI.set(model, param, value)
@@ -237,8 +240,6 @@ function solve_scs_moi(data::PMedianData; max_iters)
               (MOI.RawParameter("acceleration_lookback"), 0)]
     @timeit "SCS MOI" solve_moi(data, SCS.Optimizer, params=params)
 end
-
-
 
 function run_benchmark(;num_facilities, num_customers, num_locations,
         time_limit_sec, max_iters)